+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Technical - Upgrade ESLint rules.
 - Technical - Ensure that all files follow the ESLint rules.
 
+### Fixed
+- Technical - Exclude from pre-commit-hook linter files that are not commited.
+
 ## RELEASE 5.4.0 - 2019-11-29
 ### Added
 - Smart Relationship/Collection - Expose a serializer to simplify the serialization.